Check to see that the OpenSSL session is setup correctly. Fixes LM-88.
If the SSL Context is not setup correctly the call to SSL_new will fail as well
causing SSL_set_bio to segfault. Added a check to see that the context is not
NULL and that the session is created properly.

Fix for LM1.2 by Mikael Hallendal, forward-ported to LM1.3 by Senko Rasic.