Protect EAI_NODATA with #ifdef. Fixes LM-120.
A recent change to GNU libc changes the EAI_NODATA constant in netdb.h to only be
defined when _GNU_SOURCE is defined.
A google for "loudmouth EAI_NODATA" indicates that EAI_NODATA is also problematical
on FreeBSD.

I'm attaching a patch that protects the usage of EAI_NODATA with an ifdef ... I think this
is preferable to the sledgehammer of _GNU_SOURCE; while GNU libc can apparently
still return this error value, using the default message in this rare corner case should not
be a problem.