Changed so that normal XMPP "pings" are always sent when keep alive is enabled.
The change to disable the periodic sending of spaces in order to ensure traffic over the socket made the LM clients disconnect periodically with specific gateways.

The implementation looks correct but it might be the firewall software that is not handling it correctly. This commit reverts the behaviour to always send the spaces even if the platform supports the TCP keep alive settings.