Worked on the LmSocket interface and added a LmSoupSocket class.
LmSoupSocket will be a possibly temporary implementation of the LmSocket
interface on top of the SoupSocket API in order to get something going quickly.
Depending on how it works out it might be made the default implementation.

Added signals and worked a bit to model the LmSocket API after the SoupSocket
API as it seems quite similar to what I had in mind.