Added LmResolver interface.
The LmResolver interface will be implemented in one blocking fashion and one
implementation that uses Libasyncns. Designing this as an interface will make
it easy to implement an asynchronous version using platform specific
technologies for Windows and Mac OS X.

It will also make it easier to create a mock implementation to simulate various
error codes to test the layers using the resolver.