Changed so that we create a resolver with various arguments depending on type.
Instead of having the lookup_host and lookup_srv virtual methods there is now
only one lookup method and the resolver is created using new_for_host or
new_for_srv instead.

The host or domain/srv values are set as properties on the base class for the
sub classes to read out.